organised the file structure
