+++ src/community/community.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { CreateCommunityController } from './create-community/create-community.controller';
-import { CreateCommunityService } from './create-community/create-community.service';
+import { CreateCommunityController } from './create-community.controller';
+import { CreateCommunityService } from './create-community.service';
 
 @Module({
   controllers: [CreateCommunityController],
