+++ src/community/create-community.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { CreateCommunityDto } from 'src/dto/create-community.dto';
+import { CreateCommunityDto } from 'src/community/dto/create-community.dto';
 
 
 @Injectable()
