+++ src/product/product.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { CreateProductController } from './create-product/create-product.controller';
-import { CreateProductService } from './create-product/create-product.service';
+import { CreateProductController } from './create-product.controller';
+import { CreateProductService } from './create-product.service';
 
 @Module({
   controllers: [CreateProductController],
